Refactor @Controller -> @Service
