+++ src/main/java/de/adesso/Jekyll2cmsApplication.java
@@ -1,6 +1,6 @@
 package de.adesso;
 
-import de.adesso.controller.CmdController;
+import de.adesso.service.CmdService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
@@ -11,7 +11,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class Jekyll2cmsApplication implements ApplicationRunner {
 
 	@Autowired
-	private CmdController cmdController;
+	private CmdService cmdService;
 
 	public static void main(String[] args) {
 		SpringApplication.run(Jekyll2cmsApplication.class, args);
@@ -19,8 +19,8 @@ public class Jekyll2cmsApplication implements ApplicationRunner {
 
 	@Override
 	public void run(ApplicationArguments args) throws Exception {
-		cmdController.init(args);
-		cmdController.parse();
-		cmdController.execute();
+		cmdService.init(args);
+		cmdService.parse();
+		cmdService.execute();
 	}
 }
